item_edit and register and login

+++ admin/brand.php
@@ -62,7 +62,7 @@ include 'include/header.php';
 							<td>
 
 								 <a href="brand_edit.php?id=<?php echo $brand['id'] ?>" class="btn btn-outline-warning btn-sm">Edit</a>
-								  <a href="brand_delete.php?id=<?php echo $brand['id'] ?>" class="btn btn-outline-danger btn-sm">Delete</a></td>
+								  <a onclick="javaScript: return confirm('Are You Sure You Want To Delete')" href="brand_delete.php?id=<?php echo $brand['id'] ?>" class="btn btn-outline-danger btn-sm">Delete</a></td>
 
 						</tr>
 					<?php } ?>
